+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/store/BrowserStore.kt
@@ -25,7 +25,8 @@ class BrowserStore(
 ) : Store<BrowserState, BrowserAction>(
     initialState,
     BrowserStateReducer::reduce,
-    middleware
+    middleware,
+    "BrowserStore"
 ) {
     init {
         initialState.selectedTabId?.let {