+++ mobile/android/android-components/components/feature/customtabs/src/main/java/mozilla/components/feature/customtabs/store/CustomTabsServiceStore.kt
@@ -10,5 +10,6 @@ class CustomTabsServiceStore(
     initialState: CustomTabsServiceState = CustomTabsServiceState()
 ) : Store<CustomTabsServiceState, CustomTabsAction>(
     initialState,
-    CustomTabsServiceStateReducer::reduce
+    CustomTabsServiceStateReducer::reduce,
+    threadNamePrefix = "CustomTabsService"
 )