+++ mobile/android/android-components/components/lib/state/src/main/java/mozilla/components/lib/state/internal/StoreThreadFactory.kt
@@ -4,21 +4,32 @@
 
 package mozilla.components.lib.state.internal
 
+import mozilla.components.lib.state.Store
+import mozilla.components.support.base.utils.NamedThreadFactory
 import java.util.concurrent.Executors
 import java.util.concurrent.ThreadFactory
 
 /**
- * Custom [ThreadFactory] implementation wrapping [Executors.defaultThreadFactory] that allows
- * asserting whether a caller is on the created thread.
+ * Custom [ThreadFactory] implementation wrapping [Executors.defaultThreadFactory]/[NamedThreadFactory]
+ * that allows asserting whether a caller is on the created thread.
  *
  * For usage with [Executors.newSingleThreadExecutor]: Only the last created thread is kept and
  * compared when [assertOnThread] is called.
+ *
+ * @param threadNamePrefix Optional prefix with which to name threads for the [Store]. If not provided,
+ * the naming scheme will be deferred to [Executors.defaultThreadFactory]
  */
-internal class StoreThreadFactory : ThreadFactory {
+internal class StoreThreadFactory(
+    threadNamePrefix: String?
+) : ThreadFactory {
     @Volatile
     private var thread: Thread? = null
 
-    private val actualFactory = Executors.defaultThreadFactory()
+    private val actualFactory = if (threadNamePrefix != null) {
+        NamedThreadFactory(threadNamePrefix)
+    } else {
+        Executors.defaultThreadFactory()
+    }
 
     override fun newThread(r: Runnable): Thread {
         return actualFactory.newThread(r).also {